イベントハンドラを分離
